srf02: move to PX4Rangefinder and cleanup
